# Tracking nodes with pods with affinity