**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:

Adds a list to nodeinfo snapshot that tracks the nodes with at least one pod with affinity terms. This is useful to avoid iterating over unrelated nodes for inter-pod affinity priority and inter-pod anti-affinity predicate

**Which issue(s) this PR fixes**:
Fixes #

**Special notes for your reviewer**:

This will help avoid the performance degradation from #84669

Before:
```
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           6481403 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           8203782 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           7180056 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000          13674520 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000          12901412 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000          11295372 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          12800015 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000          11432830 ns/op

```
After:
```
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           5593549 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           7663664 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           6269434 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000          12713162 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000          11840169 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000          10337815 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          12088693 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000          10787583 ns/op
```

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```


/assign @Huang-Wei